% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocorrelation_bound_functions.R
\name{autocorrelation_coeff_h}
\alias{autocorrelation_coeff_h}
\title{Computes the approximate functional autocorrelation coefficient at a given lag.}
\usage{
autocorrelation_coeff_h(f_data, lag)
}
\arguments{
\item{f_data}{the functional data matrix with observed functions in the columns}

\item{lag}{the lag to use to compute the single lag test statistic}
}
\value{
numeric value; the approximate functional autocorrelation coefficient at lag h.
}
\description{
Computes the approximate functional autocorrelation coefficient at a given lag.
}
