% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocorrelation_bound_functions.R
\name{B_h_bound}
\alias{B_h_bound}
\title{Compute weak white noise confidence bound for autocorrelation coefficient.}
\usage{
B_h_bound(f_data, lag, alpha = 0.05, M = NULL)
}
\arguments{
\item{f_data}{the functional data matrix with observed functions in the columns}

\item{lag}{the lag to use to compute the single lag test statistic}

\item{alpha}{the significance level to be used in the hypothesis test}

\item{M}{Number of samples to take when applying a Monte-Carlo approximation}
}
\value{
numeric value; the 1-alpha confidence bound for the functional autocorrelation
coefficient at lag h under a weak white noise assumption.
}
\description{
`B_h_bound` Computes an approximate asymptotic upper 1-alpha confidence bound for the functional
autocorrelation coefficient at lag h under a weak white noise assumption.
}
