% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocorrelation_bound_functions.R
\name{B_iid_bound}
\alias{B_iid_bound}
\title{Compute strong white noise confidence bound for autocorrelation coefficient.}
\usage{
B_iid_bound(f_data, alpha = 0.05)
}
\arguments{
\item{f_data}{the functional data matrix with observed functions in the columns}

\item{alpha}{the significance level to be used in the hypothesis test}
}
\value{
Numeric value; the 1-alpha confidence bound for the functional autocorrelation coefficient
at lag h under a strong white noise assumption.
}
\description{
\code{B_iid_bound} Computes an approximate asymptotic upper 1-alpha confidence bound for the functional
autocorrelation coefficient at lag h under the assumption that f_data forms a strong white noise
}
