% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_kernels.R
\name{bartlett_kernel}
\alias{bartlett_kernel}
\title{Bartlett Kernel Function}
\usage{
bartlett_kernel(x)
}
\arguments{
\item{x}{the point value at which the kernel function is evaluated}
}
\value{
A scalar value; the value of the Bartlett kernel function at the point value x.
}
\description{
\code{bartlett_kernel} Computes the Bartlett kernel function at a given point value.
}
