% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{center}
\alias{center}
\title{Center functional data}
\usage{
center(f_data)
}
\arguments{
\item{f_data}{the functional data matrix with observed functions in the columns}
}
\value{
A matrix of the same form as f_data containing the centered functional data.
}
\description{
\code{center} Centers the given functional data
}
