% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_functions.R
\name{covariance_diag_store}
\alias{covariance_diag_store}
\title{List storage of diagonal covariances.}
\usage{
covariance_diag_store(f_data, K)
}
\arguments{
\item{f_data}{the functional data matrix with observed functions in the columns}

\item{K}{the range of lags 1:K to use}
}
\value{
A list containing K 2-dimensional arrays containing the diagonal covariance matrices of the
functional data, for lags h in the range 1:K.
}
\description{
\code{covariance_diag_store} Creates a list storage of approximate diagonal covariances computed
by the function diagonal_covariance_i
}
