% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_kernels.R
\name{daniell_kernel}
\alias{daniell_kernel}
\title{Daniell Kernel Function}
\usage{
daniell_kernel(x)
}
\arguments{
\item{x}{the point value at which the kernel function is evaluated}
}
\value{
A scalar value; the value of the Daniell kernel function at the point value x.
}
\description{
\code{daniell_kernel} Computes the Daniell kernel function at a given point value.
}
