% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocorrelation_bound_functions.R
\name{diagonal_autocov_approx_0}
\alias{diagonal_autocov_approx_0}
\title{Compute the diagonal covariance}
\usage{
diagonal_autocov_approx_0(f_data)
}
\arguments{
\item{f_data}{the functional data matrix with observed functions in the columns}
}
\value{
A numeric value; integral approximation of the diagonal covariance of the functional data.
}
\description{
\code{diagonal_autocov_approx_0} Computes the diagonal covariance of the given functional data.
}
