% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_functions.R
\name{diagonal_covariance_i}
\alias{diagonal_covariance_i}
\title{Compute the approximate diagonal covariance matrix for lag windows defined by i}
\usage{
diagonal_covariance_i(f_data, i)
}
\arguments{
\item{f_data}{the functional data matrix with observed functions in the columns}

\item{i}{the index in 1:T that we are computing the covariance for}
}
\value{
A 2-dimensional array, encoding the covariance matrix of the functional data for lag
windows defined by i.
}
\description{
\code{diagonal_covariance_i} Computes the approximate diagonal covariance matrix of the functional
data for lag windows defined by i.
}
