% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simulations.R
\name{fgarch_1_1}
\alias{fgarch_1_1}
\title{\code{fgarch_1_1} Simulates an fGARCH(1,1) process with N independent observations, each observed}
\usage{
fgarch_1_1(N, J, delta = 0.01, burn_in = 50)
}
\arguments{
\item{N}{the number of fGARCH(1,1) curves to sample.}

\item{J}{the number of points at which each curve is sampled (the resolution of the data).}

\item{delta}{a parameter used in the variance recursion of the model.}

\item{burn_in}{the number of initial samples to burn (discard).}
}
\value{
A list containing two J x N matrices, the former containing the sample of fGARCH(1,1)
curves and the latter containing the respective variance values.
}
\description{
\code{fgarch_1_1} Simulates an fGARCH(1,1) process with N independent observations, each observed
}
\examples{
f <- fgarch_1_1(100, 50)

}
