% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_functions.R
\name{iid_covariance}
\alias{iid_covariance}
\title{Compute part of the covariance under a strong white noise assumption}
\usage{
iid_covariance(f_data)
}
\arguments{
\item{f_data}{the functional data matrix with observed functions in the columns}
}
\value{
A 2-dimensional matrix containing one of the two independent sums in the computation of the
covariance.
}
\description{
\code{iid_covariance} A helper function used to compute one of the two independent sum terms in the
computation of the approximate covariance of the functional data under a strong white noise assumption.
}
