% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_tests.R
\name{independence_test}
\alias{independence_test}
\title{\code{independence_test} Performs a test for independence and identical distribution of functional
observations. The test relies on a dimensional reduction via a projection of the data on the K
most important functional pricnipal components.}
\usage{
independence_test(f_data, components, lag, alpha = 0.05,
  suppress_raw_output = FALSE, suppress_print_output = FALSE)
}
\arguments{
\item{f_data}{the functional data matrix with observed functions in the columns}

\item{components}{the number of principal components to project the data on}

\item{lag}{the maximum lag to include; this can be seen as the bandwidth.}

\item{alpha}{the significance level to be used for the hypothesis test}

\item{suppress_raw_output}{Boolean value, FALSE by default. If TRUE, the function will not return the list
contining the p-value, quantile, and statistic.}

\item{suppress_print_output}{Boolean value, FALSE by default. If TRUE, the function will not print any
output to the console.}
}
\value{
A list containing the statistic and p-value of the independence test.
}
\description{
\code{independence_test} Performs a test for independence and identical distribution of functional
observations. The test relies on a dimensional reduction via a projection of the data on the K
most important functional pricnipal components.
}
\details{
The "independence" portmanteau test is a test of independence and identical distributuion based on a
dimensionality reduction by projecting the data onto the most important functional principal components.
It is based on the resulting lagged cross-variances. This test is not for general white noise series, and
may not hold under functional conditionally heteroscedastic assumptions. Please consult the vignette for a
deeper exposition, and consult the reference for a complete treatment.
}
\examples{
b <- brown_motion(250, 100)
independence_test(b, components = 3, lag = 5)

}
\references{
Gabrys R., & Kokoszka P. (2007). Portmanteau Test of Independence for Functional Observations.
Journal of the American Statistical Association, 102:480, 1338-1348, DOI: 10.1198/016214507000001111.
}
