% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_kernels.R
\name{parzen_kernel}
\alias{parzen_kernel}
\title{Parzen Kernel Function}
\usage{
parzen_kernel(x)
}
\arguments{
\item{x}{the point value at which the kernel function is evaluated}
}
\value{
A scalar value; the value of the Parzen kernel function at the point value x.
}
\description{
\code{parzen_kernel} Computes the Parzen kernel function at a given point value.
}
