% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_tests.R
\name{spectral_test}
\alias{spectral_test}
\title{\code{spectral_test} Computes the p-value of the size alpha spectral density based white noise
test.}
\usage{
spectral_test(f_data, kernel = "Bartlett", bandwidth = "adaptive",
  alpha = 0.05, suppress_raw_output = FALSE,
  suppress_print_output = FALSE)
}
\arguments{
\item{f_data}{the functional data matrix with observed functions in the columns}

\item{kernel}{the kernel function to use. The currently supported kernels are 'Bartlett', 'Parzen',
and 'Daniell'. The default kernel is 'Bartlett'.}

\item{bandwidth}{specifies the bandwidth to use. Currently admitted arguments are positive
integers, 'static' which computes the bandwith p via p = n^(1/(2q+1)) where n is the sample
size and q is the kernel order, or 'adaptive' which uses a bandwith selection method that
is based on the functional data.}

\item{alpha}{= the significance level to be used for the test; 0.05 by default.}

\item{suppress_raw_output}{Boolean value, FALSE by default. If TRUE, the function will not return the list
contining the p-value, quantile, and statistic.}

\item{suppress_print_output}{Boolean value, FALSE by default. If TRUE, the function will not print any
output to the console.}
}
\value{
A list containing the p-value, the quantile, and a boolean value indicating whether or not the
hypothesis is rejected.
}
\description{
The "spectral" portmanteau test is based on the spectral density operator. It essentially measures
the proximity of a functional time series to a white noise - the constant spectral density operator of an
uncorrelated series. Unlike the "single-lag" and "multi-lag" tests, this test is not for general white noise
series, and may not hold under functional conditionally heterscedastic assumptions.
}
\examples{
b = brown_motion(100, 50)
spectral_test(b)
spectral_test(b, kernel = 'Parzen', bandwidth = 'adaptive')
spectral_test(b, kernel = 'Daniell', bandwidth = 2)

}
\references{
Characiejus V., & Rice G. (2019). A general white noise test based on kernel lag-window estimates of the
spectral density operator. Econometrics and Statistics, submitted.
}
